requirement.txt
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.7.5
django-filter==0.9.2
django-guardian==1.2.5
django-toolbelt==0.0.1
djangorestframework==3.1.1
gunicorn==19.2.1
Markdown==2.6.1
pep8==1.6.2
psycopg2==2.6
selenium==2.45.0
six==1.9.0
static3==0.5.1
